% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dend.R
\name{dend}
\alias{dend}
\title{: Dendogram}
\usage{
dend(x)
}
\arguments{
\item{x}{data frame}
}
\description{
Groups the samples in a dendogram
This function will group samples in a dendogram so that we can verify how many clonal groups are present in the data.
}
\author{
Dayana
}
