% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uni_gt.R
\name{uni_gt}
\alias{uni_gt}
\title{: Unique clone}
\usage{
uni_gt(data)
}
\arguments{
\item{data}{data frame}
}
\description{
Find the unique genotypes in the data set
This function will condense the unique data rows to see how many clones we have
}
